Initial cut of move away from single-page app.
